Extract Qt GUI code from main.py into separate interface/desktop dir
